pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "echpressure2"
version = "0.0.0"
dependencies = ["numpy", "scipy", "pyyaml"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
